Properly generates array with random value now
readloop:
using r0 as value of I but then reloading it to be testPrompt caused it
to print an adress. I'm guessing of i

The program ran properly but I noticed in gdb with command x/10 &array
it listed everything as 0. I added a scanf to catch and sure enough, the
correct values are being populated.
I guess after the program returns properly?(doesnt throw a 0 in gdb) you
lose everything in you array

I also noticed it gives address
0x21038	35  1	25  52
0x21048	2   0	0   0
0x21058 0   0

0x21058 - 0x21038 == 0x20 == 32
but we have 40 bytes set aside.  32 + 2 more 4 byte chunks = 40