Massive Debug Session
Code review with A.K.,
we saw 5 outputs, there for we were only printing i
values kept getting smaller due to incrementing the adress by 1 byte
A.K. pointed out the decreasing numbers as an example of Endianess, as
the bytes get moved to the left

explained the pushing and popping registers as way to return

went though everything line by line tracing every bit of logic. The code
is inherently confusing and when I noticed that a few commits ago...I
probably should've restarted.

Here we go!